backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
tensorflow>=2.16.0
numpy>=1.26.0
Pillow>=10.0.0
pydantic>=2.0.0
